agente-juridico-breviliere/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
langgraph>=0.2.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
redis>=5.0.0
chromadb>=0.4.0
pydantic-settings>=2.2.0
structlog>=24.1.0
httpx>=0.27.0
openai>=1.12.0
google-generativeai>=0.4.0
python-docx>=1.1.0
weasyprint>=61.0
alembic>=1.13.0
docxtpl>=0.16.0
